requirements.txt
Flask==1.1.2
Flask_Cors==3.0.8
flask_marshmallow==0.13.0
Flask_SQLAlchemy==2.4.3
earthengine_api==0.1.225
click==7.1.2
ee==0.4
unicode_slugify==0.1.3
